refactor(pci): remove dependency on controller variables
 ref: MANAGER-9538
