+++ packages/manager/modules/pci/src/projects/project/additional-ips/order/order.controller.js
@@ -17,7 +17,7 @@ import {
   TRACKING_PREFIX_FORM_SUBMIT,
   TRACKING_GUIDE_LINKS,
 } from './order.constants';
-import { setDefaultSelections } from '../../setStepperDefaults.utils';
+import { setDefaultSelections } from '../../../../components/project/stepper-defaults-selection/stepper-defaults-selection.utils';
 
 export default class AdditionalIpController {
   /* @ngInject */
@@ -68,7 +68,7 @@ export default class AdditionalIpController {
     };
     this.loadMessages();
     this.initIp();
-    setDefaultSelections(this, DEFAULTS_MODEL, this.loadingDefaultValues);
+    this.setDefaults();
   }
 
   static getMaximumQuantity(product) {
@@ -394,5 +394,14 @@ export default class AdditionalIpController {
 
   onGatewayModelSelection(gateway) {
     this.selectedGatewaySize = gateway;
+  }
+
+  setDefaults() {
+    this.loadingDefaultValues = true;
+    setDefaultSelections(this, DEFAULTS_MODEL, 'currentStep').finally(() => {
+      this.$timeout(() => {
+        this.loadingDefaultValues = false;
+      });
+    });
   }
 }
